PointCloud refactor
The point here is to hide the actual structure of the point cloud data,
specifically PointData and the std::vector<PointData>.
The data can be accessed with an AttribData API and a raw void* ptr.
This allows us to load the raw data directly into an OpenGL buffer, and
contains the data we need to create interleaved vertex attributes into that buffer,
while still allowing us to iterate over each element, if necessary, using ForEachAttrib().

In the case of PointCloud this is probably overkill, but I want to use this same API for
GaussianCloud in the future. Because GaussianCloud can optionally have some or all spherical
harmonic coefficients, we don't know the data layout at compile time. This AttribData API
gives us the ability to only include what we need at run-time.