Added --nosh option, to skip loading/rendering of spherical harmonics data
This reduces the memory requirements for large scenes, and can give
a performance boost as well.

GaussianCloud will use BaseGaussianData or FullGaussianData depending on this option.
This will intern reduce the gpu memory required for each vertex.
The shaders also use the FULL_SH pre-processor symbol define to handle this option.